# documentation/api/openapi.yaml
openapi: 3.1.0
info:
  title: my.smallbra.in
  description: |-
    These are the APIs for the my.smallbra.in website based on the [OpenAPI 3.1.0](https://spec.openapis.org/oas/v3.1.0) specification. 
  version: 1.0.0
  contact:
    email: danatolman@icloud.com
  # license:
  #   name: Apache 2.0
  #   url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /artifacts:
    get:
      operationId: getPortfolioArtifacts
      description: Get all portfolio artifacts' data
      summary: Get all artifacts' data
      parameters:
        - $ref: '#/components/Parameters/portfolio/page'
      tags:
        - portfolio
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/portfolio/GetPortfolioArtifacts/200'
        '416':
          description: Range Not Satisfiable, no resource or page 𝑛 found.
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/416'
              example:
                message: '416 Range Not Satisfiable: No resource or page 12 found'
      deprecated: false

    post:
      operationId: createPortfolioArtifact
      description: Create a new portfolio artifact
      summary: Create new artifact
      tags:
        - portfolio
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/portfolio/CreatePortfolioArtifact/201'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Missing Fields:
                  $ref: '#/components/Responses/shared/4xx-examples/IncludeAllFields'
                No Category Provided:
                  $ref: '#/components/Responses/shared/4xx-examples/NoCategoryProvided'
                Missing Closing Curly Brace:
                  $ref: '#/components/Responses/shared/4xx-examples/MissingClosingCurlyBrace'
                Expected Double-quoted Property Name:
                  $ref: '#/components/Responses/shared/4xx-examples/ExpectedDoubleQuotedPropertyName'
                Unexpected Token:
                  $ref: '#/components/Responses/shared/4xx-examples/UnexpectedToken'
                Bad Control Character:
                  $ref: '#/components/Responses/shared/4xx-examples/BadControlCharacter'
        '401':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              example:
                message: 'Not authorized'
                stack: 'Error: Not authorized\n at ...'
        '403':
          description: Not authorized as admin
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Not Authorized as Admin:
                  $ref: '#/components/Responses/shared/4xx-examples/NotAuthorizedAsAdmin'
      requestBody:
        $ref: '#/components/RequestBodies/CreateArtifact'
      deprecated: false

    head:
      operationId: getPortfolioArtifactHead
      description: Get headers only for portfolio artifacts request at this endpoint
      summary: Get headers only
      tags:
        - portfolio
      responses:
        '204':
          description: No content
          headers:
            Access-Control-Allow-Credentials:
              $ref: '#/components/Headers/security/Access-Control-Allow-Credentials'
            Access-Control-Allow-Origin:
              $ref: '#/components/Headers/security/Access-Control-Allow-Origin'
            Content-Security-Policy:
              $ref: '#/components/Headers/security/Content-Security-Policy'
            Cross-Origin-Embedder-Policy:
              $ref: '#/components/Headers/security/Cross-Origin-Embedder-Policy'
            Cross-Origin-Opener-Policy:
              $ref: '#/components/Headers/security/Cross-Origin-Opener-Policy'
            Cross-Origin-Resource-Policy:
              $ref: '#/components/Headers/security/Cross-Origin-Resource-Policy'
            Referrer-Policy:
              $ref: '#/components/Headers/security/Referrer-Policy'
            Strict-Transport-Security:
              $ref: '#/components/Headers/security/Strict-Transport-Security'
            X-Content-Type-Options:
              $ref: '#/components/Headers/security/X-Content-Type-Options'
            X-Download-Options:
              $ref: '#/components/Headers/security/X-Download-Options'
            X-Frame-Options:
              $ref: '#/components/Headers/security/X-Frame-Options'
            X-Permitted-Cross-Domain-Policies:
              $ref: '#/components/Headers/security/X-Permitted-Cross-Domain-Policies'
            X-XSS-Protection:
              $ref: '#/components/Headers/security/X-XSS-Protection'
            Connection:
              $ref: '#/components/Headers/connection/Connection'
            Keep-Alive:
              $ref: '#/components/Headers/connection/Keep-Alive'
            Content-Length:
              $ref: '#/components/Headers/metadata/Content-Length'
            Content-Type:
              $ref: '#/components/Headers/metadata/Content-Type'
            Date:
              $ref: '#/components/Headers/metadata/Date'
            Origin-Agent-Cluster:
              $ref: '#/components/Headers/optimization/Origin-Agent-Cluster'
            Vary:
              $ref: '#/components/Headers/optimization/Vary'
      deprecated: false
  /artifacts/{artifactId}:
    get:
      operationId: getArtifactById
      description: Get individual artifact by target ID
      summary: Get specific artifact data
      tags:
        - portfolio
      parameters:
        - $ref: '#/components/Parameters/portfolio/artifactId'
      responses:
        '200':
          description: Successful response

          content:
            application/json:
              schema:
                $ref: '#/components/Responses/portfolio/GetPortfolioArtifactById/200'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Category Not Found:
                  $ref: '#/components/Responses/shared/4xx-examples/ArtifactNotFound'
                Resource Not Found:
                  $ref: '#/components/Responses/shared/4xx-examples/ResourceNotFound'
      deprecated: false

    patch:
      operationId: patchPortfolioArtifact
      description: Update an existing portfolio artifact by Id
      summary: Update an existing portfolio artifact
      tags:
        - portfolio
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/Parameters/portfolio/artifactId'
      responses:
        '200':
          description: Update successful

          content:
            application/json:
              schema:
                $ref: '#/components/Responses/portfolio/PatchPortfolioArtifactById/200'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Missing Closing Curly Brace:
                  $ref: '#/components/Responses/shared/4xx-examples/MissingClosingCurlyBrace'
                Expected Double-quoted Property Name:
                  $ref: '#/components/Responses/shared/4xx-examples/ExpectedDoubleQuotedPropertyName'
                Unexpected Token:
                  $ref: '#/components/Responses/shared/4xx-examples/UnexpectedToken'
                Bad Control Character:
                  $ref: '#/components/Responses/shared/4xx-examples/BadControlCharacter'
        '401':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              example:
                message: 'Not authorized'
                stack: 'Error: Not authorized\n at ...'
        '403':
          description: Not authorized as admin
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Not Authorized as Admin:
                  $ref: '#/components/Responses/shared/4xx-examples/NotAuthorizedAsAdmin'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Category Not Found:
                  $ref: '#/components/Responses/shared/4xx-examples/ArtifactNotFound'
                Resource Not Found:
                  $ref: '#/components/Responses/shared/4xx-examples/ResourceNotFound'

      requestBody:
        $ref: '#/components/RequestBodies/PatchArtifact'
      deprecated: false

    delete:
      operationId: deletePortfolioArtifact
      description: Delete an existing portfolio artifact by Id
      summary: Delete an existing portfolio artifact
      tags:
        - portfolio
      security:
              - bearerAuth: []
      parameters:
        - $ref: '#/components/Parameters/portfolio/artifactId'
      responses:
        '200':
          description: Delete successful

          content:
            application/json:
              schema:
                $ref: '#/components/Responses/portfolio/DeletePortfolioArtifactById/200'
              example:
                message: Artifact removed
        '401':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Not Authorized:
                  $ref: '#/components/Responses/shared/4xx-examples/NotAuthorized'
        '403':
          description: Not authorized as admin
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Not Authorized as Admin:
                  $ref: '#/components/Responses/shared/4xx-examples/NotAuthorizedAsAdmin'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/Responses/shared/4xx'
              examples:
                Category Not Found:
                  $ref: '#/components/Responses/shared/4xx-examples/ArtifactNotFound'
                Resource Not Found:
                  $ref: '#/components/Responses/shared/4xx-examples/ResourceNotFound'
      deprecated: false
...
servers:
  - url: 'http://localhost:4500/api'
components:
  securitySchemes:
    bearerAuth:
      description: JSON Web Token provision for Authorized HTTP requests
      type: http
      scheme: bearer
      bearerFormat: JWT
  Schemas:
...
    CreateArtifact:
      description: Schematic for object for creating a new portfolio artifact.
      type: object
      required:
        - title
        - author
        - category
        - visible
      properties:
        title:
          type: string
          description: Title of the post
          minLength: 1
          example: Swagger Test Portfolio Artifact
        description:
          type: string
          description: Description of portfolio artifact
          minLength: 1
          example: This is a description
        author:
          type: string
          description: Author's name
          minLength: 1
          example: Dana Tolman
        category:
          type: string
          description: Category of the portfolio artifact
          minLength: 1
          example: swagger
        url:
          type: string
          description: URL to the portfolio artifact. Can be a preview or a live display.
          minLength: 7
          example: https://www.example.com
        repository:
          type: string
          description: URL to the repository that hoolds the portfolio artifact, e.g. a GitHub repo.
          minLength: 7
          example: https://www.github.com/user/repo-name
        visible:
          type: boolean
          description: Visibility status of the portfolio artifact
          default: true
          example: true
        previewImage:
          type: string
          description: A URL to an image file which will serve as a preview image for the artifact.
          example: "http://localhost:8081/storage/images/smallbrain-logo3.png"
        tags:
          type: array
          description: Can be a single tag or a list of tags.
          items:
            type: string
          example: ['swagger 1', 'swagger 2', 'swagger 3']
        slug:
          type: string
          description: Custom slug for the artifact. If empty, will auto-generate from the title.
          minLength: 1
          required: false
          example: swagger-portfolio-artifact
    PatchArtifact:
      description: Schematic for object for updating a portfolio artifact.
      type: object
      properties:
        title:
          type: string
          description: Title of the portfolio artifact
          minLength: 1
          example: Swagger Portfolio Artifact Patched
        description:
          type: string
          description: Description of portfolio artifact
          minLength: 1
          example: A portfolio artifact testing out swagger.
        author:
          type: string
          description: Author's name
          minLength: 1
          example: John Doe
        category:
          type: string
          description:  Category of the portfolio artifact
          minLength: 1
          example: swagger
        url:
          type: string
          description: URL to the portfolio artifact. Can be a preview or live display
          minLength: 7
          example: https://example.com
        repository:
          type: string
          description: URL to the repository that holds the portfolio artifact.
          minLength: 7
          example: https://www.github.com/user/repo-name
        visible:
          type: boolean
          description: Visibility status of the portfolio artifact
          example: true
        previewImage:
          type: string
          description: A URL to an image file which will serve as a preview image for the artifact
          example: "https://source.unsplash.com/random/300x200"
        slug:
          type: string
          description: "Slug path for the file. Should appear at '/portfolio/<slug>'. If none is provided, then the slug will be auto-generated."
          example: swagger-test-title
          required: false
        tags:
          type: array
          description: if more than one, must be presented as comma-separated values
          items:
            type: string
          example: ['tag 1', 'tag 2', 'tag 3']
...
  RequestBodies:
...
    CreateArtifact:
      type: object
      required: true
      content:
        multipart/form-data:
          schema:
            $ref: '#/components/Schemas/CreateArtifact'
        application/json:
          schema:
            $ref: '#/components/Schemas/CreateArtifact'
      description: Object to create new portfolio artifact
    PatchArtifact:
      type: object
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/Schemas/PatchArtifact'
      description: Object to update portfolio artifact
...